feat: edit authentication method in POST token to meet new ClientRegistration requirements
