+++ src/oneid/oneid-common/utils/src/main/java/it/pagopa/oneid/common/utils/ClientConstants.java
@@ -1,8 +1,9 @@
-package it.pagopa.oneid.service.utils;
+package it.pagopa.oneid.common.utils;
 
 public abstract class ClientConstants {
 
   public static String ENCRYPTION_ALGORITHM = "AES";
+  public static String HASHING_ALGORITHM = "SHA-256";
   public static int CLIENT_SECRET_BIT_LENGTH = 256;
   public static int ACS_INDEX_DEFAULT_VALUE = 0;
